This patch includes 1)Changed all gpatohva to copy_from_gpa and copy_to_gpa 2)copying intr_ctx info into irq_desc

+++ hypervisor/arch/x86/guest/vmcall.c
@@ -183,7 +183,7 @@ int vmcall_vmexit_handler(struct vcpu *vcpu)
 
 #ifdef HV_DEBUG
 	case HC_PROFILING_OPS:
-		ret = hcall_profiling_ops(vm, param1, param2);
+		ret = (int32_t)hcall_profiling_ops(vm, param1, param2);
 		break;
 #endif
